backend/requirements.txt
# EDITH AR Backend — Python Requirements
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
websockets==13.1
pydantic==2.9.0
python-dotenv==1.0.1
aiofiles==24.1.0
# TTS
elevenlabs==2.45.0
gTTS==2.5.3
# Advanced features
chromadb>=0.5.0
numpy>=1.24.0
scipy>=1.11.0
Pillow>=10.0.0
qrcode[pil]>=7.4.2
